getLabels移植
