+++ src/main/java/click/divichart/controller/YearlyDividendController.java
@@ -11,6 +11,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import java.util.List;
+
 /**
  * 年別配当グラフ用コントローラ
  */
@@ -34,7 +36,12 @@ public class YearlyDividendController {
     public String index(Model model, @AuthenticationPrincipal UserDetails user) {
         log.debug("年別配当グラフ表示");
 
-        String labels = service.getLabels(NUM_OF_YEARS);
+        // new start
+        List<Integer> pastYears = service.getPastYears(NUM_OF_YEARS);
+        String labels = service.getLabels(pastYears);
+        // new end
+
+        //String labels = service.getLabels(NUM_OF_YEARS);
         String chartData = service.getChartData(NUM_OF_YEARS, user.getUsername());
 
         YearlyDividendDto yearlyDividendDto = new YearlyDividendDto(
