+++ src/test/java/click/divichart/service/DividendIncreaseServiceTest.java
@@ -42,37 +42,6 @@ class DividendIncreaseServiceTest {
     }
 
     @Test
-    void testGetLabels_withMultipleYears() {
-        List<Integer> pastYears = Arrays.asList(2021, 2022, 2023);
-        String expected = "\"2021年\",\"2022年\",\"2023年\"";
-        assertEquals(expected, dividendIncreaseService.getLabels(pastYears));
-    }
-
-    @Test
-    void testGetLabels_withSingleYear() {
-        List<Integer> pastYears = Collections.singletonList(2023);
-        String expected = "\"2023年\"";
-        assertEquals(expected, dividendIncreaseService.getLabels(pastYears));
-    }
-
-    @Test
-    void testGetLabels_withEmptyList_shouldThrowException() {
-        List<Integer> pastYears = Collections.emptyList();
-        Exception exception = assertThrows(
-                IllegalArgumentException.class, () -> dividendIncreaseService.getLabels(pastYears)
-        );
-        assertEquals("過去の年のリストが空です", exception.getMessage());
-    }
-
-    @Test
-    void testGetLabels_withNullList_shouldThrowException() {
-        Exception exception = assertThrows(
-                IllegalArgumentException.class, () -> dividendIncreaseService.getLabels(null)
-        );
-        assertEquals("過去の年のリストが空です", exception.getMessage());
-    }
-
-    @Test
     void testGetDividendIncreaseData() {
         // テストデータを準備
         List<Integer> years = Arrays.asList(2022, 2023);
